fix: planning-v2 refresh project sidebar
Planning v2 is a SPA, no refresh of the page is made when a milestone is
created or edited. This causes issue in the sidebar since it cannot
reflects what the user is currently doing: a new current milestone is
created but does not appear in the sidebar unless the page is refreshed.

In order to fix that, the proposal is the following: once a milestone
has been added or edited, a REST call is made by pv2 to retrieve the
(new) project sidebar config and pv2 asks to the sidebar to reflect the
new config.

The issue is that is that reactivity of the sidebar config has been
broken by the bump vue: 3.2.37 -> 3.3.4[0] This commit fixes this
reactivity issue by making sure that the injection type of the config
matches what is returned by `unserializeConfiguration` (`Configuration |
undefined`). This means that even if they are not loaded, inner
components have to deal that the config might be `undefined`.

To test, go to backlog of a ALM/Scrum project:
* create a past release => nothing changes
* create a future release => nothing changes
* create a current release => the milestone appears in the sidebar
* The same with sprint sub milestones.

[0]: git #tuleap/stable/2380a33b21e3d2415a512fdb08e2fa2e23312d1c

Change-Id: I5942f08a817cb5206f1721a9406828bb636182ec