Fix request #35495 Do not allow suspended or deleted user in REST SVN notifications
When setting an SVN notification using the PUT or POST route, administrator
can set suspended or deleted users in notification, which is wrong.

An error must be raised if a non alive user is set.

How to test
-----------

* Using the SVN PUT/POST REST route, define a notification with at least
  one suspended or deleted user

=> An error is raised

* Using the SVN PUT/POST REST route, define a notification with only
  active or restricted users

=> No error, notification is saved

Note: another commit will be done to add REST automated tests.
This is done in a dedicated commit because we need to do some setups to
introduce suspended user in REST tests.

Change-Id: I1c32d35b3515d3d5d7d24a4e0758fe09365d329e