Use @tuleap/project-sidebar 2.6.1 in the MediaWiki Standalone plugin
This is a quick workaround until the new version can be properly
integrated upstream.

The new sidebar file is deployed directly instead of using the patch to
make things easier. The Gerrit patch would need to be rewritten a bit to be
used.

Part of request #35492: MediaWiki should use sidebar 2.6.1

Change-Id: Ie0299ee792ad24f92f89ddc5a73aba7fd75ca26d