Closes request #35813: Remove the use of the Code Coverage plugin in the nightly pipeline
The Code Coverage plugin is deprecated (and has been removed from our CI environment).
The Coverage plugin does not seem to support the merge of Cobertura
reports so we just store the HTML reports instead like it is done for
the PHP code.

You can test the generation of the reports with `COLLECT_COVERAGE=true pnpm test`

Change-Id: Ie38a8e87d38896d5999c355dc045dee81e071f3c