chore(svn): Build REST SVN Settings in a dedicated object
This refactoring aims to remove the double validations of data done for
Settings comming to REST data.

In this first step, we move the building of the object in a dedicated
unit tested class. In a second step, we will move all the checks done in
SettingsRepresentationValidator directly in the builder to remove the
double verifications.

This is a followup of request #35495 Do not allow suspended or deleted user in REST SVN notifications

No functionnal change expected.

Change-Id: Id37e57895a9527564c98c0364dde6548070487a2