feat: Create forge__cce user
This user will be used in custom code execution context.

Later we will have to create either a `PFUSER::CCE_ID` constant or a new
class CustomCodeExecutionUser to expose the reserved id 70 of this user.

The user is visible in site admin » users » suspended.

Part of story #35093: execute custom code as artifact post action

Change-Id: I04b23c2b5086b77308cfff9c5e52e1276f09ff0a