feat: ActionsRunner dispatch an event to collect more PostCreationTask
Part of story #35093 execute custom code as artifact post action

ActionsRunner cannot know if plugin TrackerCCE is installed or not and
then if it can trigger CustomCodeExecutionTask. To fix that,
ActionsRunner launch an event that collect a list PostCreationTask and
then run them at the end of workflow.

*Testing:*

Create/Update an artifact, a debug entry should appear in log files :

`CustomCodeExecutionTask called on artifact #id for changeset #id`

Change-Id: I1db5411ad65e1fbaf62cce68c2dccd2ce2416e07