chore(svn): Remove SVN Core usage of updateSVNAccess
The removed call corresponds to the situation where an action should
have kicked the update of .SVNAccessFile of the SVNCore repository. As
there are no longer SVNCore repositories created and SVN Plugin takes
care of the ones that were migrated, this code is dead.

It's also part of the work on svn 1.10+ support because calls to
updateSVNaccess must be updated in order to support Repositories instead
of the bunch of params.

Last, it also fixes a bug on LDAP login update. When a user changed
their login, the .SVNAccessFile of the projects they were member of was
no longer updated.

Part of: request #29976 Remove SVN core
Part of: request #31541 svn 1.10+ no longer supports duplicate sections

Change-Id: Id560c794700fe6c63680832a8f78e9c648bbc2f9