chore(👋 svn-core): Remove commit-email.pl
And there is goes for the whole (in)famous perl stack.

Note: for perl dependencies in the RPM specfile it's mostly a guess as
I'm not really sure there is no hidden / side effect dependency that
could be covered here.

With removal of perl-libwww-perl and perl-LWP-Protocol-https (no longer
needed after hudson.pl removal) there are side-effects on gitolite
(data/dumper was not an explicit dependency).

Part of: request #29976 Remove SVN core

Change-Id: Id0d3f5e3fa59d3f1884a8271e1d45486edc2d793