Allow to pass multiple mount points (or none) to the executed WASM module
No functional change is expected but this should bring flexibility
for upcoming contributions.

To test you need to build and redeploy the Wasmtime wrapper. The
easiest way to achieve that is to re-build and redeploy the RPM in your
dev env (`nix-build src/additional-packages/tuleap-wasmtime-wrapper-lib.nix`).

Part of story #35093 execute custom code as artifact post action

Change-Id: I7a456bbb52a5dbebcc218eba2953f55710b73443