Drop unnecessary getter in SearchCriteria
`public readonly` does the same thing with less boilerplate.

Note that is likely that we will be able to directly pass
`Option<StatusCriterion>` in the constructor and to make it work as
expected with the latest version of Valinor.

part of story #35821 Pull request new home page - Filter on Authors

Change-Id: Ifa0ba8b59f84a57a953dd8ccd0b9776d06a4c0a9