Switch to PHP 8.2 by default
Dev instances are already running PHP 8.2, this contribution moves
the frontend of the production instances to it.

PHP 8.1 dependencies are kept for now to make possible to quickly
rollback in case of a major issue.

Some old and deprecated internal nginx documentation has been removed.

Part of request #30344: Run Tuleap with PHP 8.2

Change-Id: Iead5d91c3bf35ef98a1ffa1506ff0ae4454769cf