feat: Introduce duck-typed field
How to test:
- Enable the feature flag:
tuleap config-set feature_flag_cross_tracker_search_duck_typed_fields 1
- When you type a field name that does not exist in any tracker, you
will get an error telling you that we don't know what the name
references. Previously, it resulted in error 500.
- No other functional change.

Notes:
When building the SQL query, we will need to know the "type" of a given
"duck-typed field" so that we are able to write the query for a given
comparison. For example, given "initial_effort = 5", we need to know
that "initial_effort" is a numeric like "int or float" so that we can
query the corresponding tracker changeset value DB tables.
Having an enum (instead of a class) will allow us to use a match
expression on it. We did not have enums for the type of metadata, so we
had to use a switch, but the goal is the same.

Using DuckTypedField, I can build the "type" of a duck-typed field and
it can be used both for raising errors to the end-user, and later to
build the actual SQL query.

part of story #10710 search on fields with duck typing

Change-Id: I723615a027d6ad0f723823d07ec824eaaeaecbff