feat: init filters dropdown
part of story #35821 Pull request new home page - Filter on Authors

This contribution introduces a new lib
@tuleap/plugin-pullrequest-selectors-dropdown. It will contain the new
graphical element whose role is to provide a way to create filters
through a list of attributes + an autcompleter.

Since there is a chance for this pattern to be democratized in Tuleap, I
chose to do the implementation inside a local library, and as a custom
element using hybrids. This way, it should not be hard to reuse the day
we want to have such selectors elsewhere because we'll "only" have to move
it to tuleap/lib/frontend, and do a bit of renaming in the package.json.

How to test:
- `make post-checkout`
- Go to the pull-requests homepage of a repository and add a parameter
  `tab=homepage` in the query string of its url
--> There is a [+ Add filter] button above the list of pull-requests
--> A dropdown opens if you click on it
--> There is only one entry "Author"
--> Nothing happens if you click it

Change-Id: I109333c7d34344023d745dfef35788a67c4d8259