feat: ability to remove a module
Now an empty state is displayed when a module has not been uploaded or
has been removed.

The tracker administrator can remove a module via the web interface
without having to `rm` the file on the filesystem.

Prepare the field for the next changeset (table to display logs, switch
to activate the module, …).

The main interest of this contribution is the addition of the empty
state and the ability to remove a module. Design is subject to change
and copywriting will be enhanced in dedicated contribution.

Part of story #35093: execute custom code as artifact post action

Change-Id: Id101512b3e99529f66198d21bfb6d08b77c26b65