chore(svn/el9): Extract write of .SVNAccessFile from BackendSVN
In order to better control how .SVNAccessFile is managed, we can no
longer rely on the historical way to write this file down on the
filesystem with BackendSVN because we need to be able to call it from
different contexts.

Even more important, we change the logic associated to .SVNAccessFile
writing. Until now, there were two parts: the so called "DefaultBlock"
was supposed to be Project wide because it defines the user groups
(that doesn't change from one repository to another in a given project)
as well as the default permissions (that doesn't change either because
it's defined by project privacy). In the future, the default permission
block will change from one repository to another so the write of the
whole .SVNAccessFile is now associated to the Repository object it
belongs to.

Part of: request #31541 svn 1.10+ no longer supports duplicate sections

Change-Id: Ia051b233f454fb9c6faaee4b430d5b183916399d