feat: system check in error if nb backend workers is 0
Tracker CCE plugin needs to have at least 1 backend worker. With the
recent update of `sys_nb_backend_workers` config management, we reduce
the risk to have a configuration not aligned with plugins usage. In
order to cover all bad situations, we now terminate the system check
event in error as soon as the config is not valid.

To test this contribution, remove the entry `sys_nb_backend_workers` from `forgeconfig` table and manually set it to 0 via `/etc/tuleap/conf/local.inc`: `$sys_nb_backend_workers = 0;`. This is the scenario where existing users don't read deployment guide and keep the value in the local.inc file.

Process manually the system check event:
```
tuleap queue-system-check
tuleap process-system-events default
```

THe event is in error. If you disable plugins that depends on backend
workers (tracker cce, program, …) and you relaunch the system check, it
does not end up in error anymore (at least not for the current subject ^^).

If you increase the number of backend workers, enable back plugins, and
relaunch the systen check event, it does not end up in error either.

Part of story #35093: execute custom code as artifact post action

Change-Id: I52b17174909f333ae11ad7b10a538153780549d2