refactor: RetrieveUsedFields takes tracker id into account
No functional change expected. CI should be happy.

Notes:
We need to retrieve fields from several different trackers for
cross-tracker search with duck-typed fields. Previously, I would have
added a method like `withSuccessiveFields()` in the stub, but it makes
the tests dependent on the order of execution. Instead, we can change
the stub so it behaves more realistically and returns the fields that
have the correct tracker ID. Tests will now need to set up the fields
with the right tracker IDs.
As a bonus, we now only pass this interface once in
DryRunDuckTypingFieldCollector.

part of story #10710 search on fields with duck typing

Change-Id: Ifccc5c1dc1f45af510041b5e898370d1f11c494e