READ permissions must be checked on field
part of story #10710 search on fields with duck typing

How to test:
Remove read permission on searched field:
 - the permission is silently ignored (to not leak field name)
 - the field is not taken in account for search

Why?
Without this change, if a user have access to an artifact, he can retrieve artifact data based on a field he can not read, it leaks data. This should be not a production issue has the feature is hide bellow a feature flag

Change-Id: Ic091c1bbacd1d40e33528570c1d45ade0fc0a92a