feat: display CCE logs
Each time the CCE module is executed, a log entry is recorded.

Those logs are now displayed in the administration interface.
Administrators have the ability to display the details of the execution to
see the payloads and copy them in their clipboard.

In case of error, the error message is displayed, if any.

For the sake of readability, the json payloads are pretty printed and
syntax highlighted.

We should probably mention somewhere that only the last 50 entries are
kept, but don't know how yet, so it will be for a next contribution.

Part of story #35093: execute custom code as artifact post action

Change-Id: Iefefa8e80527769d4ce4182d920bfa826bb5e24f