feat: Tracker admin receive a mail when WASM fail
Part of story #35093 execute custom code as artifact post action

To inform tracker admin that its uploaded WASM module fail on an
artifact update, we send an email with changeset id and artifact id. The
admin can then go to wasm logs to check details

*Testing:*

In a tracker with a WASM module, if execution is successful, nothing
change. But if execution failed, the tracker administrator (only)
receive a mail.

Change-Id: I1da46d52babb8d0b6c595b9c2eee42658f3631c7