feat(svn/el9): introduce ability to disable default permissions
This cannot be tested manually at the moment. This patch introduce the
concept of repository that can have default permission or not as well as
the consequences for generation of the default block.

The cache mecanism is a sensitive part of the patch as it was introduced
on larger platform with lot's of svn repositories and lot's of user
groups with a lot's of users (yeah). We can no longer cache the whole
default block, however it's still possible to cache the groups
definition (once per project) as well as the "plugin override" as it's
project based as well.

Part of: request #31541 svn 1.10+ no longer supports duplicate sections

Change-Id: I6f59046ae3dcc7b9f8d7d4d9b8c1dd22f4d385df