fix: Output CCE post action should be strictly validated
In b41a03d161277799b1d603dded459414bc7755f6 the annotations have not
been adjusted to parseable by Valinor. As a result the output of the CCE
post action is too loosely validated (e.g. it is possible with a
direction "wrong").

Part of story #35093: execute custom code as artifact post action

Change-Id: I5220de18d803f928268b3b5745cbb822fe910832