chore: Remove ArtifactUserCanViewStub usages
With recent contributions to ArtifactTestBuilder[0], we can now remove
ArtifactUserCanViewStub in order to not have two different ways for
doing things.

No functional changes, tests must pass.

Note: in order to improve readability of the StoredReverseLinkTest, the
setUp and build method has been inlined. The size of those two methods
does not justify the added complexity (forcing us to move back and
forth to understand what each test was doing).

Related to request #35862

Change-Id: I388a97ce50ee469edae8bcea338680abd01922d9