feat: Delete logs when removing module
Part of story #35093 execute custom code as artifact post action

When removing a module, the logs are not needed anymore. And if the Tracker
admin upload a new module, the logs of previous module are not relevant.

*Testing:*

In a Tracker with some logs, delete your uploaded module, then upload a
new one. No logs should be displayed. Logs of others Trackers should not
be impacted.

Change-Id: I803e8bab6c3ca2dc2ef14207ec8f2a3364513232