feat(svn/el9): allow to disable default permissions
It's possible to deactivate default '[/]' block via web ui and REST API.

XML import will be covered in a dedicated commit.

A potential issue was caught during review: it would have been more
accurate to associate "has default permissions" to the svnaccessfile
history rather than the repository itself. When an admin select an old
version of .SVNAccessFile, they can expect that the result will be
exactly the permissions that were set back in time. However, with the
default permission toggle at repository level it's not possible.

However, this issue was caught late in the development and the use of
"use old version" is supposed to be low. In order to not block the
ability to migrate to el9 for subversion users, we decided the following
tradeoff:
- The use of default permission is recorded in Project History (so we
  know the status of previous version of .SVNAccessFile)
- It's no longer possible to save the used version, the svn admin only
  have the information to compare, it reduces the risk of misuse.

How to test
-----------

In GUI you can untick `Use default permissions` then only `[groups]`
block is generated by Tuleap.

In REST you can create / update repositories with or without `[/]` as
well as specific permissions.

Part of: request #31541 svn 1.10+ no longer supports duplicate sections

Change-Id: Idb2d251244987d5efdc68290bb717b1323562834