fix: Lock rows when creating/updating indexed items
Without the lock there is a possibility that we create 2 times the same
item which will conflict later on.

The issue is hard to reproduce because it requires to have 2 (or more)
simultaneous queries creating the same item. The insertion part of the
simultaneous queries needs to occur after the point where check if the
item is already present or not.

Closes request #35872: Possible indexing issues when quickly updating the same artifact field

Change-Id: I73e50145ffc413aad886fd8577d3c198e5fe17a5