feat: Query for Lesser than comparison for numeric fields
How to test:
- Enable the feature flag:
tuleap config-set feature_flag_cross_tracker_search_duck_typed_fields 1
- Set up two or three trackers with int or float fields. The fields must
have the same shortname, for example "initial_effort".
- You can write a TQL query to search on initial_effort with lesser than
`<` and a numeric value
- Writing a TQL query like `initial_effort < ''` is not rejected. The
search does not yield results but should be raised as a query error. It
will be done on a dedicated commit.
- You can search on multiple fields and combine results with "AND" /
"OR".

Notes:
I also added tests for "OR" expressions. The goal is to verify that we
correctly added the table aliases on the SQL query. With this test, if I
remove the hash from a table alias, the test will fail due to non-unique
table SQL error.

part of story #10710 search on fields with duck typing

Change-Id: I087fef1af1c44d1da2d9e369fa6708f07abd3eff