feat: Basic e2e cover of Tuleap Functions for Tracker
The test is based on post-action-add-comment that was introduced in
Tuleap sources to develop the feature. The binary has been compiled from
plugins/tracker_cce/wasm_modules/post-action-add-comment and is added as
a fixture.

For now the test only cover the following:
* upload of a function
* check function is executed after artifact creation

Other features are considered not worth to be covered because
meaningless (like check admin logs) or redundant (like artifact update).
This can be revised later.

Part of story #35093: execute custom code as artifact post action

Change-Id: Ie336f57a28bef6537905b00d8a44ce82b75426f4