refactor: Flatten comparison matching
No functional change for cross-tracker TQL queries on numeric duck-typed
fields.

Why?
Due to the structure of the TQL language, we are forced to identify the
type of the Comparison before we know its left-hand part (the
Searchable). It makes things painful for us because for each type of
comparison, we will repeat the code to build a duck-typed field and
match its type. Having more flexibility could help us group related
things together: it is the same process to retrieve a duck-typed field,
no matter the comparison.

Using an enum breaks the "visitor" spirit, but there are also
advantages:
- match expression raises an error if a branch is not handled, so we
should have an equivalent guarantee of completeness as visitor
- we can pass different parameters to each branch. With visitor, we are
forced to pass the same parameters. This limitation is the reason for
the generic VisitorParameter and why we must pass so much stuff in
nested collection parameters.
- We are no longer forced to instantiate a different class for each
comparison. The architecture should be easier to understand (no jumping
between interfaces and visitors).

part of story #10710 search on fields with duck typing

Change-Id: I9e1c0c08201cb1032fdb755526528384036a9b77