feat: Query for Greater Than Or Equal comparison for numeric fields
How to test:
- Enable the feature flag:
tuleap config-set feature_flag_cross_tracker_search_duck_typed_fields 1
- Set up two or three trackers with int or float fields. The fields must
have the same shortname, for example "initial_effort".
- You can write a TQL query to search on initial_effort with greater
than or equal `>=` and a numeric value
- Writing a TQL query like `initial_effort >= ''` is not rejected. The
search always yields all artifacts but should be raised as a query
error. It will be done on a dedicated commit.
- You can search on multiple fields and combine results with "AND" /
"OR".

Change-Id: I60f8905a1df2e8d82705d0c2d5e5f68d14e709d3