fix(artifacts): Move artifacts must not decrease artifact deletion
This is part of request #34032 Move of artifacts should not depend on artifact deletion

The move action is no more regulated using the artifact deletion limit
(this was confusant for users). Instead, each tracker can define if move
action is allowed or not (allowed by default). The artifact deletion
limit must not be decreased by moving an artifact.

Change-Id: I467376ed5853489abe5be05f664015145a03e54e