refactor: selection is always an array in Lazybox
This contribution aims to bring consitency in the way selected values
are passed to the caller code.

Before this patch:
- single lazybox were returning either null, either a value
- multiple lazybox were returning either an empty array, either an array
  filled with values.

Now, both will return an array, empty or filled with values. There will
be only one value in the array for single lazyboxes, and 1+ for multiple
lazyboxes.

implements request #35893 Lazybox selection should always be an array

-- no functional change expected

Change-Id: I0ed27d74a635c6b3dce4c0318727bdcc3d85fc40