fix: forge__tracker_workflow_manager should not trigger WASM pipeline
Closes request #36482 Updates from technical user should not trigger Tuleap Functions for Trackers

When a transition rule is applied, the update is performed by
forge__tracker_workflow_manager. This update trigger the WASM pipeline
which can override the original update and so the transition rule is
again applied, and again (infinity loop).

To avoid that, we just filter the trigger by the user.

Change-Id: I16fd6f242a8b642c7860532d9dfcb61a89d9ac06