refactor: timetracking overview wiget uses Pinia
part of request #35894 Convert timetracking-overview-widget to Vue3 + TS

In this contribution, vuex is replaced by pinia.

It was tricky because several instances of the overview widget can
exist on the same dashboard. Hence, it is needed to have separated store
instances, in order to avoid conflicts between the different widgets.

-- no functional change expected

Change-Id: Ide45ff2a6410956a1a3c491732aac82fd0caf153