fix(document): Close request #36801 Progress bar in Document upload is stuck to 0%
When uploading one or multiple files in Document, then the upload progress bar always displays 0% and then 100% when finished.

Technical note: The FakeItem was previsouly defined in the store action.
Now, using Vue3, defining this in the action will define this FakeItem
as readonly. We have to define it in the Vue file to be well modified
and taken into account by Vue components.

Change-Id: I9ece30fec3fad3f388671ffde3b96f8045a49ade