feat: Add a "Labels" entry in add-filter dropdown
part of story #35825 Pull request new home page - Filter on labels

This contribution adds a "Labels" entry in the [+ Add filter] dropdown.
When clicked, it shows an autocompleter for project labels.

Note:

The selection is not handled yet. It will be done in a dedicated commit.

How to test:
- Open the [+ Add filter] dropdown
--> There is a "Labels" item in the menu

- Click on it
--> The side panel opens
--> It displays an autocompleter showing the labels defined in the
current project

- Click on "Author"
--> The author autocompleter is now displayed

Change-Id: I2db4de6288322c84f422c3066be2f5dd2b4a97f8