feat: Display new PR homepage by default
part of story #36795 Show/hide closed pull-requests

Now that the new PR homepage view covers the same functional scope, if
not better, than the angular pull-requests dashboard, it is time to
display it by default.

The old view is kept behind a feature flag, so users will be able to put
it back as default if needed.

How to test:
- Apply this patch
- Go to the pull-request homepage by clicking on the tab "Pull Requests"
--> The new homepage is displayed

- Now exec `tuleap config-set feature_flag_display_old_pull_requests_dashboard 1`
- Go back to the pull-request homepage by clicking twice on the tab
  "Pull Requests" (urls need to be regenerated backend-side)
--> The old view is displayed

Change-Id: Ie63f96190de952809c1b3ff8d74acf8e9847c792