comparisons between int and float should be valid in regular tql
part of story #10710 search on fields with duck typing

How to test:
in a tracker have an artifact ith an inteeger field with value 10
create a TQL query `integer_field > 5.5`

without this patch:
 - an error is displayed, telling you you can't compare int to float

with this patch
 - 10 is bigger than 5.5 so your artifact is displayed by query

Why?
Implementing cross tracker search on duck typed field has shown us an inconsistence
In widget in order to ease search you can compare `numeric` field.
If you have an initial_effort in a release tracker  who is a int, an an intial_effort
in sprint who is a float, querying `intial_effort > 0` will retrieve all artifact or release and sprint.

As the constraint have been removed in widget, we must adapt behaviour in tracker report too.

Change-Id: I2c097c5f8f2d6c4944b1b78b4c81edfa5981be88