refacto: stop using the same representation for PUT and GET REST route
Part of story #36485 have custom services to /projects/:id/project_services

Using the same representation as a return of REST GET route and as
a parameter in another PUT REST route can lead to functional inconsistency (and some headaches).
In the PUT route, only `is_enabled` key from the ServiceRepresentation is supported and the usage of others keys are  ckecked.
However, some keys in ServiceRepresentation has been added after and, the usage of new keys have not been checked in the `PUT` route (which should be a PATCH route btw). Even if it works in the end the "task fail successfully".
This contribution intend to avoid future inconsistency by splitting the PUT representation and the GET representation.

No functionnal change expected

Change-Id: Ia7ebb64d99ab6af2ff42287f810c50be5a225f39