fix: Have a better empty state message
part of story #35825 Pull request new home page - Filter on labels

Currently, when all the pull-requests are closed, the empty state
message says "No result match your query". This is a bit misleading
since it can be displayed while no filters have been defined yet.

The proposition is the following:
- When some filters are defined -> display a "No matching results"
  message.

- When the closed pull-requests are hidden -> display a "No open
  pull-requests" message

Please note that the case "There are no open nor closed pull-requests"
does not need to be implemented, since it already is. When no
pull-requests exist in the reporitory, the backend serves a static page
containing a dedicated empty state + the create pull-request button.

How to test:
- Go to a repository where there are only closed pull-requests
--> Empty state says "no open pull-request"

- Add some filters
--> Empty state says "No matching results"

Change-Id: Ie13f9f77bd14abd6e0ec2146aac705a686be0fb1